+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: bdevessi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/30 14:57:05 by bdevessi          #+#    #+#             */
-/*   Updated: 2018/12/10 16:06:45 by bdevessi         ###   ########.fr       */
+/*   Updated: 2018/12/14 11:10:05 by bdevessi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -163,7 +163,7 @@ char	*pathjoin(char *s1, char *s2)
 	return (str - i - j);
 }
 
-size_t		nb_len(int n)
+size_t		nb_len(uint64_t n)
 {
 	size_t	i;
